% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{feed.extract}
\alias{feed.extract}
\title{Extract RSS/Atom feed}
\usage{
feed.extract(url)
}
\arguments{
\item{url}{URL for the feed.}
}
\value{
A list containing the following elements:

- title: Title of the original site.

- link: A link to the original site.

- updated: When the feed was last updated.

- items: A data frame with records for each entry in the feed.
}
\description{
Read feed metadata and entries.
}
\examples{
feed.extract("https://feeds.feedburner.com/RBloggers")
feed.extract("http://journal.r-project.org/rss.atom")
}

