% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{parse.rdf}
\alias{parse.rdf}
\title{Parse feeds encoded in RDF/XML.}
\usage{
parse.rdf(feed)
}
\description{
Parse feeds encoded in RDF/XML.
}
\examples{
parse.rdf(feed.read("http://feeds.feedburner.com/oatmealfeed"))
}
\references{
https://en.wikipedia.org/wiki/RDF/XML
}

